backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - MySQL
aiomysql==0.2.0
pymysql==1.1.0

# Vector DB and embeddings
faiss-cpu==1.7.4
sentence-transformers>=2.3.0
numpy>=1.24.3
huggingface-hub>=0.24.0,<1.0

# Groq API
groq==0.4.1

# Redis for async event processing
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
